+++ src/views/Grants/GrantsHome.vue
@@ -108,11 +108,7 @@
 </template>
 
 <script>
-const addOneDayToDate = function (date) {
-  const newDate = new Date(date);
-  newDate.setDate(newDate.getDate() + 1);
-  return newDate;
-};
+import { isPastLastDay } from "@/utils/ended";
 // eslint-disable-next-line no-unused-vars
 import { attachInternalLinks, attachSearchEvents } from "@/utils/dom.js";
 
@@ -198,14 +194,14 @@ export default {
     filterGrants(status) {
       if (status === "current") {
         this.filteredAndSortedGrants = _.filter(this.allGrants, (grant) => {
-          if (new Date(addOneDayToDate(grant.end)) > new Date()) {
+          if (!isPastLastDay(grant.end)) {
             return grant;
           }
         });
       }
       if (status === "expired") {
         this.filteredAndSortedGrants = _.filter(this.allGrants, (grant) => {
-          if (new Date(addOneDayToDate(grant.end)) < new Date()) {
+          if (isPastLastDay(grant.end)) {
             return grant;
           }
         });